+++ core/src/main/resources/phoenix/create_tables.sql
@@ -8,9 +8,16 @@ CREATE TABLE IF NOT EXISTS topology_components ("id" BIGINT NOT NULL, "name" VAR
 CREATE TABLE IF NOT EXISTS topology_editor_metadata ("topologyId" BIGINT NOT NULL, "data" VARCHAR, "timestamp"  BIGINT, CONSTRAINT pk PRIMARY KEY ("topologyId"))
 CREATE TABLE IF NOT EXISTS tag ("id" BIGINT NOT NULL, "name" VARCHAR(256), "description" VARCHAR(256), "timestamp" BIGINT, CONSTRAINT pk PRIMARY KEY ("id"))
 CREATE TABLE IF NOT EXISTS tag_storable_mapping ("tagId" BIGINT NOT NULL, "storableNamespace" VARCHAR(32) NOT NULL, "storableId" BIGINT NOT NULL, CONSTRAINT pk PRIMARY KEY ("tagId", "storableNamespace", "storableId"))
-CREATE TABLE IF NOT EXISTS streaminfo ("id" BIGINT NOT NULL, "streamId" VARCHAR(256), "fieldsData" VARCHAR, "timestamp" BIGINT, CONSTRAINT pk PRIMARY KEY ("id"))
 CREATE TABLE IF NOT EXISTS notifierinfos ("id" BIGINT  NOT NULL, "name" VARCHAR, "jarFileName" VARCHAR, "className" VARCHAR, "timestamp"  BIGINT, "properties" VARCHAR, "fieldValues" VARCHAR, CONSTRAINT pk PRIMARY KEY ("id"))
-CREATE TABLE IF NOT EXISTS sequence_table ("id" VARCHAR, "datasources" BIGINT, "datafeeds" BIGINT, "parser_info" BIGINT, "files" BIGINT, "topologies" BIGINT, "topology_components" BIGINT, "tag" BIGINT,  "streaminfo" BIGINT, "notifierinfos" BIGINT, CONSTRAINT pk PRIMARY KEY ("id"))
+CREATE TABLE IF NOT EXISTS streaminfo ("id" BIGINT NOT NULL, "topologyId" BIGINT, "streamId" VARCHAR(256), "fieldsData" VARCHAR, "timestamp" BIGINT, CONSTRAINT pk PRIMARY KEY ("id"))
+CREATE TABLE IF NOT EXISTS topology_components ("id" BIGINT NOT NULL, "topologyId" BIGINT, "name" VARCHAR, "type" VARCHAR, "configData" VARCHAR, CONSTRAINT pk PRIMARY KEY ("id"))
+CREATE TABLE IF NOT EXISTS topology_sources ("id" BIGINT NOT NULL, "topologyId" BIGINT, "name" VARCHAR, "type" VARCHAR, "configData" VARCHAR, CONSTRAINT pk PRIMARY KEY ("id"))
+CREATE TABLE IF NOT EXISTS topology_source_stream_mapping ("sourceId" BIGINT NOT NULL, "streamId" BIGINT NOT NULL, CONSTRAINT pk PRIMARY KEY ("sourceId", "streamId"))
+CREATE TABLE IF NOT EXISTS topology_sinks ("id" BIGINT NOT NULL, "topologyId" BIGINT, "name" VARCHAR, "type" VARCHAR, "configData" VARCHAR, CONSTRAINT pk PRIMARY KEY ("id"))
+CREATE TABLE IF NOT EXISTS topology_processors ("id" BIGINT NOT NULL, "topologyId" BIGINT, "name" VARCHAR, "type" VARCHAR, "configData" VARCHAR, CONSTRAINT pk PRIMARY KEY ("id"))
+CREATE TABLE IF NOT EXISTS topology_processor_stream_mapping ("processorId" BIGINT NOT NULL, "streamId" BIGINT NOT NULL, CONSTRAINT pk PRIMARY KEY ("processorId", "streamId"))
+CREATE TABLE IF NOT EXISTS topology_edges ("id" BIGINT NOT NULL, "topologyId" BIGINT, "fromId" BIGINT NOT NULL, "toId" BIGINT NOT NULL, "streamGroupingsData" VARCHAR, CONSTRAINT pk PRIMARY KEY ("id"))
+CREATE TABLE IF NOT EXISTS sequence_table ("id" VARCHAR, "datasources" BIGINT, "datafeeds" BIGINT, "parser_info" BIGINT, "files" BIGINT, "topologies" BIGINT, "topology_components" BIGINT, "tag" BIGINT,  "streaminfo" BIGINT, "notifierinfos" BIGINT, "topology_sources" BIGINT, "topology_sinks" BIGINT, "topology_processors" BIGINT, "topology_edges" BIGINT, CONSTRAINT pk PRIMARY KEY ("id"))
 CREATE SEQUENCE IF NOT EXISTS datasources_sequence
 CREATE SEQUENCE IF NOT EXISTS datafeeds_sequence
 CREATE SEQUENCE IF NOT EXISTS parser_info_sequence
@@ -20,3 +27,8 @@ CREATE SEQUENCE IF NOT EXISTS tag_sequence
 CREATE SEQUENCE IF NOT EXISTS files_sequence
 CREATE SEQUENCE IF NOT EXISTS streaminfo_sequence
 CREATE SEQUENCE IF NOT EXISTS notifierinfos_sequence
+CREATE SEQUENCE IF NOT EXISTS topology_components_sequence
+CREATE SEQUENCE IF NOT EXISTS topology_sources_sequence
+CREATE SEQUENCE IF NOT EXISTS topology_sinks_sequence
+CREATE SEQUENCE IF NOT EXISTS topology_processors_sequence
+CREATE SEQUENCE IF NOT EXISTS topology_edges_sequence